fix(e2e): resolve helpers/browsers imports from ../shared after the restructure
