+++ test/e2e/updater/updater-e2e.mjs
@@ -35,8 +35,8 @@ import {
   rmDir,
   summary,
   localConfigOverrides,
-} from './helpers.mjs';
-import {findSnapshot, findZip, extractZip, discoverFirefoxBinary, findGreDir} from './browsers.mjs';
+} from '../shared/helpers.mjs';
+import {findSnapshot, findZip, extractZip, discoverFirefoxBinary, findGreDir} from '../shared/browsers.mjs';
 
 const UPDATER_URL = 'chrome://firefox-scripts/content/ui/updater.html';
 const FORCE_UTILS_STALE = 'RDFDataSource.sys.mjs';
